Track undeleted series IDs per-shard with inmem
This commit adds a bitset into each shard's engine to be used to
track undeleted series ids. Currently inmem support is implemented.

When new series are added to the shard, the series id is added
to the bitset. When series are deleted from the shard, the series
ids are removed from the bitset.

Becasue each shard shares the same inmem index reference, the bitset
is stored in the `ShardIndex`, which is local to each shard, and then
different references are passed into the shared `Index` object, depending
on which shard is writing the series.